gcp-functions/rollup/requirements.txt
duckdb
gcsfs
google-cloud-storage
pandas
functions-framework
pytz
